x86: Avoid unnecessary overflow in sched_clock
Patchwork-id: 42877
O-Subject: [RHEL6.3 PATCH BZ 765720]: sched,
	x86: Avoid unnecessary overflow in sched_clock
Bugzilla: 765720
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Motohiro Kosaki <mkosaki@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Backport linux-2.6 commit 4cecf6d401a01d054afc1e5f605bcbfe553cb9b9

In hundreds of days, the __cycles_2_ns calculation in sched_clock
has an overflow.  cyc * per_cpu(cyc2ns, cpu) exceeds 64 bits, causing
the final value to become zero.  We can solve this without losing
any precision.

We can decompose TSC into quotient and remainder of division by the
scale factor, and then use this to convert TSC into nanoseconds.

I tried to reproduce by using WRMSR to forward the time in a macroscopic
jump in order to cause a system panic, but hit a panic/lockup/etc.,
in other parts of the system that are unrelated to the timing change.
I also had a quick email exchange with
Salman Qazi, the upstream author of this patch regarding reproducibilty in
which he told me that he (and upstream) only did sanity checks on the code.
AFAICT there is no way to test this other than to leave a system running for
~6 months (208 days).  I've reviewed the code and it seems valid, it is
upstream, and AFAICT it is stable.

Successfully compiled and boot-tested by me on a variety of AMD and Intel
systems.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>