kabi: Add missing symbols to kABI
Patchwork-id: 42890
O-Subject: [PATCH RHEL63 BZ 737276 2/2] kabi: Add missing symbols to kABI
Bugzilla: 787040
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 737276
https://bugzilla.redhat.com/show_bug.cgi?id=737276

Description:
============
Request to add following symbols to kABI, with usage explanation
from partner in:
(https://bugzilla.redhat.com/show_bug.cgi?id=737276#c7)

These symbols should be listed in the whitelist because they are
used in the modules which sytemtap creates. In addition, portability
of systemtap modules will be improved.

  alloc_vm_area
  get_fs_type
  probe_kernel_read
  probe_kernel_write
  si_meminfo
  vscnprintf

We have an outbox HBA driver which has been certified for RHEL6.
Therefore, these symbols used in the HBA driver need to be in the
whitelist.

  blk_queue_rq_timed_out
  fc_attach_transport
  fc_release_transport
  fc_remote_port_add
  fc_remote_port_delete
  fc_remote_port_rolechg
  fc_remove_host
  touch_nmi_watchdog

Systemtap can insert probes on int3 instruction. This function
can cause a problem when a module is unloaded while probes are
still inserted to it. Current version of systemtap runtime does
not care about it, therefore it needs to be cared on each systemtap
module code's side. Hitachi's tracer uses the following symbols
to avoid this problem. Therefore, these symbols need to be in the
whitelist for portability of this module.

  find_module
  module_mutex
  use_module

This symbol should be listed in the whitelist, because Hitachi's
tracer uses this symbol to get local time with tiny cost. For
troubleshooting use, it is important for tracer to get local time
to compare trace with logs. It also needs to be fast enough because
tracer is required to collect many information. To meet these
requirements, we use xtime and wall_to_monotonic variable, which are
not exported, via kallsyms_on_each_symbol. Therefore, this symbol
needs to be in the whitelist for portability of this module.

  kallsyms_on_each_symbol

Brew:
=====
https://brewweb.devel.redhat.com/taskinfo?taskID=3929867

Tested:
=======
brewed test kernel properly

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>