GFS2: Update fallocate to match upstream
Patchwork-id: 42997
O-Subject: [RHEL 6.3] GFS2: Update fallocate to match upstream (bz #732744)
Bugzilla: 732744
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

The following patch contains the upstream parts of the fallocate fix
which didn't make it into 6.2. The main change is to move the
buffer zeroing into the block map function and the introduction of
a new flag to control the zeroing.

It has been tested using the script attached to the bz as well as
a few other sanity checks.

This fixes bz #732744

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>