ftrace: Balance records when updating the hash
Patchwork-id: 43026
O-Subject: [RHEL6.3 BZ 454694 PATCH 28/32] [PATCH 28/32] ftrace: Balance
	records when updating the hash
Bugzilla: 454694
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

From: Steven Rostedt <srostedt@redhat.com>

Upstream: 41fb61c2d08107ce96a5dcb3a6289b2afd3e135c
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=454694

Whenever the hash of the ftrace_ops is updated, the record counts
must be balance. This requires disabling the records that are set
in the original hash, and then enabling the records that are set
in the updated hash.

Moving the update into ftrace_hash_move() removes the bug where the
hash was updated but the records were not, which results in ftrace
triggering a warning and disabling itself because the ftrace_ops filter
is updated while the ftrace_ops was registered, and then the failure
happens when the ftrace_ops is unregistered.

The current code will not trigger this bug, but new code will.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>