mpt2sas: Fix for system hang when discovery in progress
Patchwork-id: 43191
O-Subject: [RHEL6.3 PATCH 05/30] mpt2sas: Fix for system hang when discovery in
	progress.
Bugzilla: 736229
RH-Acked-by: Rob Evers <revers@redhat.com>

Fix for issue : While discovery is in progress, hot unplug and hot plug of
enclosure connected to the controller card is causing system to hang.

When a device is in the process of being detected at driver load time then
if it is removed, the device that is no longer present will not be added
to the list. So the code in _scsih_probe_sas() is rearranged as such so
the devices that failed to be detected are not added to the list.

Upstream-
0167ac67ff6f35bf2364f7672c8012b0cd40277f [SCSI] mpt2sas: Fix for system hang when discovery in progress

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>