mpt2sas: Fix for issue deadlock between hot plug worker threads and host reset context
Patchwork-id: 43196
O-Subject: [RHEL6.3 PATCH 09/30] mpt2sas: Fix for issue deadlock between hot
	plug worker threads and host reset context.
Bugzilla: 736229
RH-Acked-by: Rob Evers <revers@redhat.com>

This is due to driver reporting a device missing to the OS then the OS sending
a SYNC_CACHE request to driver while the IO queues are locked due to host reset.

To fix the issue, the driver will be waking up the port enable context
immediately when the driver receives the reply message, instead of waiting
on the hot plug worker threads.

Upstream-
918134efe9893629407af04adf242ee3095bea4a [SCSI] mpt2sas: Fix for deadlock between hot plug worker threads and host reset context

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>