mpt2sas: Fix for issue Port Reset is taking long time to complete while issued during creating a volume
Patchwork-id: 43198
O-Subject: [RHEL6.3 PATCH 10/30] mpt2sas: Fix for issue Port Reset is taking
	long time(around 5 mins) to complete while issued during
	creating a volume.
Bugzilla: 736229
RH-Acked-by: Rob Evers <revers@redhat.com>

This is due to the slave_configuration routine is getting called when
host reset is active, and config page reads are failing, and driver
attempts to added device with stale config data.

To fix the issue, added error checking in slave_configure to check
for configuration pages failing, and return "1" so the device  is
not configured.  The config pages are failing if raid volume is
configured while issuing a host reset, thus driver is reading stale
data and proceeding to attempt to add.  The fix is to return error
so the volume is not configured.

Upstream-
6faace2a0e418b45728bcea6d3626922cf16b14b [SCSI] mpt2sas: Fix for issue Port Reset taking long time(around 5 mins)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>