mpt2sas: Release spinlock for the raid device list before blocking it
Patchwork-id: 43210
O-Subject: [RHEL6.3 PATCH 24/30] mpt2sas: Release spinlock for the raid device
	list before blocking it.
Bugzilla: 736229
RH-Acked-by: Rob Evers <revers@redhat.com>

Added code to release the spinlock that is used to protect the
raid device list before calling a function that can block. The
blocking was causing a reschedule, and subsequently it is tried
to acquire the same lock, resulting in a panic (NMI Watchdog
detecting a CPU lockup).

Upstream-
30c43282f3d347f47f9e05199d2b14f56f3f2837 [SCSI] mpt2sas: Release spinlock for the raid device list before blocking it

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>