mpt2sas: Fix for memory allocation error for large host credits
Patchwork-id: 43215
O-Subject: [RHEL6.3 PATCH 26/30] mpt2sas: Fix for memory allocation error for
	large host credits
Bugzilla: 736229
RH-Acked-by: Rob Evers <revers@redhat.com>

The amount of memory required for tracking chain buffers is rather
large, and when the host credit count is big, memory allocation
failure occurs inside __get_free_pages.

The fix is to limit the number of chains to 100,000.  In addition,
the number of host credits is limited to 30,000 IOs. However this
limitation can be overridden this using the command line option
max_queue_depth.  The algorithm for calculating the
reply_post_queue_depth is changed so that it is equal to
(reply_free_queue_depth + 16), previously it was (reply_free_queue_depth * 2).

Upstream-
aff132d95ffe14eca96cab90597cdd010b457af7 [SCSI] mpt2sas : Fix for memory allocation error for large host credits

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>