Reserve iommu page 0
Patchwork-id: 43345
O-Subject: [PATCH RHEL6.3 BZ784088] powerpc: Reserve iommu page 0
Bugzilla: 784088
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=784088

Description:
------------
Some devices have a dma-window that starts at the address 0. This allows
DMA addresses to be mapped to this address and returned to drivers as a
valid DMA address. Some drivers may not behave well in this case, since
the address 0 is considered an error or not allocated.

The solution to avoid this kind of error from happening is reserve the
page addressed as 0 so it cannot be allocated for a DMA mapping.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
RHEL 6.2

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3976530

Upstream:
---------
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;a=commit;h=d12b524f8b2f4e45cabe8bc1501e8b967d543111

Test Status:
------------
Thadeu Lima de Souza Cascardo helped me test this.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>