raid10: Improve decision on whether to fail a device with a read error
Patchwork-id: 43366
O-Subject: [PATCH 023/122] md/raid10: Improve decision on whether to fail a
	device with a read error.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

Normally we would fail a device with a READ error.  However if doing
so causes the array to fail, it is better to leave the device
in place and just return the read error to the caller.

The current test for decide if the array will fail is overly
simplistic.
We have a function 'enough' which can tell if the array is failed or
not, so use it to guide the decision.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 700c72138938cf428c74379806886c6b017d6295)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>