raid10: simplify/reindent some loops
Patchwork-id: 43417
O-Subject: [PATCH 052/122] md/raid10: simplify/reindent some loops.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

When a loop ends with a large if, it can be neater to change the
if to invert the condition and just 'continue'.
Then the body of the if can be indented to a lower level.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 1294b9c973251a5e68b62c9b40dd914517bda675)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>