raid10: Handle read errors during recovery better
Patchwork-id: 43400
O-Subject: [PATCH 065/122] md/raid10: Handle read errors during recovery better.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

Currently when we get a read error during recovery, we simply abort
the recovery.

Instead, repeat the read in page-sized blocks.
On successful reads, write to the target.
On read errors, record a bad block on the destination,
and only if that fails do we abort the recovery.

As we now retry reads we need to know where we read from.  This was in
bi_sector but that can be changed during a read attempt.
So store the correct from_addr and to_addr in the r10_bio for later
access.

Signed-off-by: NeilBrown<neilb@suse.de>
(cherry picked from commit 5e5702898e93eee7d69b6efde109609a89a61001)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>