raid5: fix a hang on device failure
Patchwork-id: 43404
O-Subject: [PATCH 071/122] md/raid5: fix a hang on device failure.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

Waiting for a 'blocked' rdev to become unblocked in the raid5d thread
cannot work with internal metadata as it is the raid5d thread which
will clear the blocked flag.
This wasn't a problem in 3.0 and earlier as we only set the blocked
flag when external metadata was used then.
However we now set it always, so we need to be more careful.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 43220aa0f22cd3ce5b30246d50ccd696d119edea)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>