add proper write-congestion reporting to RAID1 and RAID10
Patchwork-id: 43427
O-Subject: [PATCH 102/122] md: add proper write-congestion reporting to RAID1
	and RAID10.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

RAID1 and RAID10 handle write requests by queuing them for handling by
a separate thread.  This is because when a write-intent-bitmap is
active we might need to update the bitmap first, so it is good to
queue a lot of writes, then do one big bitmap update for them all.

However writeback request devices to appear to be congested after a
while so it can make some guesstimate of throughput.  The infinite
queue defeats that (note that RAID5 has already has a finite queue so
it doesn't suffer from this problem).

So impose a limit on the number of pending write requests.  By default
it is 1024 which seems to be generally suitable.  Make it configurable
via module option just in case someone finds a regression.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry-picked with conflicts from 34db0cd60f8a1f4ab73d118a8be3797c20388223)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>