clear In_sync bit on devices added to an active array
Patchwork-id: 43429
O-Subject: [PATCH 103/122] md: clear In_sync bit on devices added to an active
	array.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

When we add a device to an active array it can be meaningful to set
the 'insync' flag.  This indicates that the device is in-sync with the
array except for locations recorded in the bitmap.
A bitmap-based recovery can then bring it completely in-sync.

Internally we move that flag to 'saved_raid_disk' but forgot to clear
In_sync like we do in add_new_disk.

So clear In_sync after moving its value to saved_raid_disk.

Reported-by: Andrei Warkentin <andreiw@vmware.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit d30519fc59c5cc2f7772fa67b16b1a2426d36c95)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>