raid5: fix bug that could result in reads from a failed device
Patchwork-id: 43442
O-Subject: [PATCH 106/122] md/raid5: fix bug that could result in reads from a
	failed device.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

This bug was introduced in 415e72d034c50520ddb7ff79e7d1792c1306f0c9
which was in 2.6.36.

There is a small window of time between when a device fails and when
it is removed from the array.  During this time we might still read
from it, but we won't write to it - so it is possible that we could
read stale data.

We didn't need the test of 'Faulty' before because the test on
In_sync is sufficient.  Since we started allowing reads from the early
part of non-In_sync devices we need a test on Faulty too.

This is suitable for any kernel from 2.6.36 onwards, though the patch
might need a bit of tweaking in 3.0 and earlier.

Cc: stable@kernel.org
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 355840e7a7e56bb2834fd3b0da64da5465f8aeaa)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>