raid10: Fix bug when activating a hot-spare
Patchwork-id: 43435
O-Subject: [PATCH 108/122] md/raid10:  Fix bug when activating a hot-spare.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

This is a fairly serious bug in RAID10.

When a RAID10 array is degraded and a hot-spare is activated, the
spare does not take up the empty slot, but rather replaces the first
working device.
This is likely to make the array non-functional.   It would normally
be possible to recover the data, but that would need care and is not
guaranteed.

This bug was introduced in commit
   2bb77736ae5dca0a189829fbb7379d43364a9dac
which first appeared in 3.1.

Cc: stable@kernel.org
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 7fcc7c8acf0fba44d19a713207af7e58267c1179)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>