ensure new badblocks are handled promptly
Patchwork-id: 43447
O-Subject: [PATCH 116/122] md: ensure new badblocks are handled promptly.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

When we mark blocks as bad we need them to be acknowledged by the
metadata handler promptly.

For an in-kernel metadata handler that was already being done.  But
for an external metadata handler we need to alert it of the change by
sending a notification through the sysfs file.  This adds that
notification.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 8bd2f0a05b361e07d48bb34398593f5f523946b3)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>