raid5: never wait for bad-block acks on failed device
Patchwork-id: 43466
O-Subject: [PATCH 117/122] md/raid5: never wait for bad-block acks on failed
	device.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

Once a device is failed we really want to completely ignore it.
It should go away soon anyway.

In particular the presence of bad blocks on it should not cause us to
block as we won't be trying to write there anyway.

So as soon as we can check if a device is Faulty, do so and pretend
that it is already gone if it is Faulty.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 9283d8c5af4cdcb809e655acdf4be368afec8b58)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>