raid5: ensure correct assessment of drives during degraded reshape
Patchwork-id: 43467
O-Subject: [PATCH 120/122] md/raid5: ensure correct assessment of drives during
	degraded reshape.
Bugzilla: 747574
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

While reshaping a degraded array (as when reshaping a RAID0 by first
converting it to a degraded RAID4) we currently get confused about
which devices are in_sync.  In most cases we get it right, but in the
region that is being reshaped we need to treat non-failed devices as
in-sync when we have the data but haven't actually written it out yet.

Reported-by: Adam Kwolek <adam.kwolek@intel.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 30d7a4836847bdb10b32c78a4879d4aebe0f193b)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>