perf report: Fix stdio event name header printing
Patchwork-id: 43494
O-Subject: [PATCH RHEL6.3 BZ784888 19/97] perf report: Fix stdio event name
	header printing
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit b53af0f235b6334563c4c9274860f7c5775c07ff
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Wed Sep 21 15:43:04 2011 -0300

description
===========
In the past we tried to avoid printing the name of the event when just
one event was found in the perf.data file, after some refactorings it
ended up not printing the event name if just one hist_entry was found in
one of the events.

Fix it by always printing the name of the event, even if just one is
found.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>