perf annotate browser: Allow navigation to called functions
Patchwork-id: 43502
O-Subject: [PATCH RHEL6.3 BZ784888 27/97] perf annotate browser: Allow
	navigation to called functions
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit 34958544b3da17e98d98d7b872c6516995ad66bb
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Wed Oct 5 19:35:54 2011 -0300

description
===========
I.e. when in the annotate TUI window, if Enter is pressed over an
assembly line with a 'callq' it will try to open another TUI window with
that symbol.

This is just a proof of concept and works only on x86_64, more work is
needed to support kernel modules, userland, other arches, etc, but
should already be useful as-is.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>