perf hists browser: Update the browser.nr_entries after the timer
Patchwork-id: 43514
O-Subject: [PATCH RHEL6.3 BZ784888 33/97] perf hists browser: Update the
	browser.nr_entries after the timer
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit be83f5ed6bc46cd89b4a102b6e341ecddf7abf91
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Thu Oct 6 10:56:05 2011 -0300

description
===========
Previously the hist_browser dealt with a static tree of entries, now it
needs to update the nr_entries in the browser after the timer runs.

A better solution will come when moving using another thread for the
collapse_resort, etc, but for now this is ok.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>